Backend/requirements.txt
arrow==1.3.0
asgiref==3.7.2
certifi==2022.6.15
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-redis==5.4.0
django-rq==2.10.1
djangorestframework==3.14.0
filelock==3.4.2
gunicorn==21.2.0
packaging==23.2
platformdirs==2.6.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
rq==1.16.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
six==1.16.0
sqlparse==0.4.4
times==0.7
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
